updated datatype from long to integer

+++ src/main/java/com/busticket/dto/request/BookingRequest.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class BookingRequest {
 
     @NotNull(message = "Trip ID is required")
-    private Long tripId;
+    private Integer tripId;
 
     @NotNull(message = "Seat number is required")
     @Min(value = 1, message = "Seat number must be at least 1")
