+++ src/main/java/com/busticket/service/interfaces/IReviewService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IReviewService {
     List<Review> findAll();
-    Review findReviewById(long id);
-    List<Review> findReviewByCustomerId(long customerId);
-    List<Review> findReviewByTripId(long reviewerId);
+    Review findReviewById(Integer id);
+    List<Review> findReviewByCustomerId(Integer customerId);
+    List<Review> findReviewByTripId(Integer reviewerId);
 }
